+++ src/egfileio.c
@@ -12,39 +12,39 @@
 #include <memory.h>
 
 /* Private helpers for this translation unit. */
-int createFileWrapper(const char *filename, int attr);
-int readFile1Wrapper(int handle, int count, int bufOffset);
-int readFile2Wrapper(int handle, int count, int bufOffset, int bufSegment);
-int writeFileAtRawWrapper(int handle, int count, int bufOffset, int bufSegment, int offsetAddend);
+int16 createFileWrapper(const char *filename, int16 attr);
+int16 readFile1Wrapper(int16 handle, int16 count, int16 bufOffset);
+int16 readFile2Wrapper(int16 handle, int16 count, int16 bufOffset, int16 bufSegment);
+int16 writeFileAtRawWrapper(int16 handle, int16 count, int16 bufOffset, int16 bufSegment, int16 offsetAddend);
 
-int openFileWrapper(const char *filename, int mode) { /* Original: OpenFile(file, attrib). Open resident file service; returns a file handle. */
+int16 openFileWrapper(const char *filename, int16 mode) { /* Original: OpenFile(file, attrib). Open resident file service; returns a file handle. */
     Log(("openFileWrapper: path=%s mode=%d", filename, mode));
     /* Thin resident-service wrapper: pass filename and open mode unchanged. */
     return openFile(filename, mode);
 }
 
 // ==== seg000:0xdd5e ====
-int createFileWrapper(const char *filename, int attr) { /* Original: CreateFile(file). Create resident file service; returns a file handle/status. */
+int16 createFileWrapper(const char *filename, int16 attr) { /* Original: CreateFile(file). Create resident file service; returns a file handle/status. */
     /* attr is kept for the wrapper ABI; the resident create service uses it. */
     return createFile(filename, attr);
 }
 
 // ==== seg000:0xdd70 ====
-void closeFileWrapper(int handle) { /* Original: CloseFile(fh). Close a resident file-service handle. */
+void closeFileWrapper(int16 handle) { /* Original: CloseFile(fh). Close a resident file-service handle. */
     closeFile(handle);
 }
 
 // ==== seg000:0xdd7e ====
-int readFile1Wrapper(int handle, int count, int bufOffset) {
+int16 readFile1Wrapper(int16 handle, int16 count, int16 bufOffset) {
     return readFile1(handle, count, bufOffset);
 }
 
 // ==== seg000:0xdd92 ====
-int readFile2Wrapper(int handle, int count, int bufOffset, int bufSegment) {
+int16 readFile2Wrapper(int16 handle, int16 count, int16 bufOffset, int16 bufSegment) {
     return readFile2(handle, count, bufOffset, bufSegment);
 }
 
 // ==== seg000:0xddaa ====
-int writeFileAtRawWrapper(int handle, int count, int bufOffset, int bufSegment, int offsetAddend) {
+int16 writeFileAtRawWrapper(int16 handle, int16 count, int16 bufOffset, int16 bufSegment, int16 offsetAddend) {
     return writeFileAtRaw(handle, count, bufOffset, bufSegment, offsetAddend);
 }
