+++ src/egframe.h
@@ -3,19 +3,19 @@
 /* public interface of egframe.c */
 #include "inttype.h"
 
-void countermeasures(int eventType);
+void countermeasures(int16 eventType);
 void resetSimObjectLocks();
 void initWeaponLoadout(void);
 void drawWeaponAmmo();
-void drawWeaponSelectMarker(int weaponIdx);
-void finalizeMission(int outcome);
-void scheduleEventCheck(int eventObjIdx, uint16 priority);
-void scheduleTimedEvent(int keyVal, int delay);
-void appendMapEvent(int eventType, int eventArg);
-void placeString(int waypointIdx);
+void drawWeaponSelectMarker(int16 weaponIdx);
+void finalizeMission(int16 outcome);
+void scheduleEventCheck(int16 eventObjIdx, uint16 priority);
+void scheduleTimedEvent(int16 keyVal, int16 delay);
+void appendMapEvent(int16 eventType, int16 eventArg);
+void placeString(int16 waypointIdx);
 void initMissionStrings();
 void moveDataFar();
-int objectToScreen(int mapX, int mapY, int16 *outScreenX, int16 *outScreenY);
-int randomRange(int);
+int16 objectToScreen(int16 mapX, int16 mapY, int16 *outScreenX, int16 *outScreenY);
+int16 randomRange(int16);
 
 #endif /* F15_SE2_EGFRAME */
