+++ src/egsphere.h
@@ -2,6 +2,6 @@
 #define F15_SE2_EGSPHERE
 /* public interface of egsphere.c */
 
-void drawProjectionSphere(int skyColor);
+void drawProjectionSphere(int16 skyColor);
 
 #endif /* F15_SE2_EGSPHERE */
